+++ Fmas12d.Mobile/src/app/interfaces/location.interface.ts
@@ -1,0 +1,8 @@
+export interface Location {
+  contactDetailId?: number;
+  contactDetailTypeName?: string;
+  latitude?: number;
+  longitude?: number;
+  postcode?: string;
+  type: string;
+}
