+++ Fmas12d.Mobile/src/app/interfaces/user-availability.interface.ts
@@ -1,19 +1,10 @@
+import { Location } from './location.interface';
+
 export interface UserAvailability {
-  contactDetailId?: number;
   end: Date | string;
   id: number;
-  isAvailable: boolean;
   location: Location;
-  postcode?: string;
   start: Date | string;
   statusId: number;
 }
 
-export interface Location {
-  contactDetailId?: number;
-  contactDetailTypeName?: string;
-  latitude?: number;
-  longitude?: number;
-  postcode?: string;
-  type: string;
-}
